Solid code refractor
